+++ packages/qwik/src/utils/getInitialFieldStore.ts
@@ -1,0 +1,48 @@
+import type {
+  FieldPath,
+  FieldStore,
+  FieldValue,
+  FieldValues,
+  InitialFieldState,
+  Maybe,
+} from '../types';
+import { isFieldDirty } from './isFieldDirty';
+
+/**
+ * Returns the initial store of a field.
+ *
+ * @param initialState The initial state.
+ *
+ * @returns The initial store.
+ */
+export function getInitialFieldStore<
+  TFieldValues extends FieldValues,
+  TFieldName extends FieldPath<TFieldValues>
+>(
+  name: TFieldName,
+  {
+    value,
+    initialValue = value,
+    error = '',
+  }: InitialFieldState<TFieldValues, TFieldName>
+): FieldStore<TFieldValues, TFieldName> {
+  const dirty = isFieldDirty(
+    initialValue as Maybe<FieldValue>,
+    value as Maybe<FieldValue>
+  );
+  return {
+    internal: {
+      initialValue,
+      startValue: initialValue,
+      validate: [],
+      elements: [],
+      consumers: [],
+    },
+    name,
+    value,
+    error,
+    active: false,
+    touched: dirty,
+    dirty,
+  };
+}